Add the word "operator" before symbolic magic methods
This perhaps helps the reader understand why they're seeing symbols in the function name.

We should probably do the same thing for the function overloads, but I'm not sure what the appropriate prefix would be. Some options are:

* `operator len(x)`, `operator operator.index(x)`
* `implements len(x)`, `implements operator.index(x)`
* `supports len(x)`, `supports operator.index(x)`

etc. Without some kind of marker, it's impossible to distinguish `len(x)` from `x.len(self=...)` in the docs.